SAM-79 Add moka sample cache
