+++ src-tauri/src/op_playback_commands.rs
@@ -9,6 +9,7 @@ use crate::playback::op_playback::{
 };
 use crate::playback_ops::merge_playback::MergePlaybackOp;
 use crate::playback_ops::sample_playback::SamplePlayableOp;
+use crate::sample_cache::SampleCacheService;
 use crate::{emit_logged, log_debug, log_info, send_channel_event};
 use rodio::{OutputStream, Sink};
 use serde::Serialize;
@@ -214,10 +215,12 @@ pub struct BuildGraphResponse {
 pub async fn op_playback_build_graph(
     request: BuildGraphRequest,
     state: State<'_, Arc<OpPlaybackState>>,
+    sample_cache: State<'_, Arc<SampleCacheService>>,
     logging_service: State<'_, Arc<Mutex<LoggingService>>>,
     on_event: Channel<OpPlaybackBuildGraphEvent>,
 ) -> Result<BuildGraphResponse, String> {
     let state = state.inner().clone();
+    let sample_cache = sample_cache.inner().clone();
     let logging_service = logging_service.inner().clone();
 
     tauri::async_runtime::spawn_blocking(move || {
@@ -269,8 +272,10 @@ pub async fn op_playback_build_graph(
                     let samples = if let Some(ref samples) = op_request.samples {
                         samples.clone()
                     } else if let Some(ref file_path) = op_request.file_path {
-                        // Load samples from file
-                        load_audio_samples(file_path, sample_rate, channels)?
+                        // Load samples from file using cache
+                        let path = std::path::PathBuf::from(file_path);
+                        let buffer = sample_cache.get_or_load(path, sample_rate, channels)?;
+                        buffer.data.clone()
                     } else {
                         return Err(format!(
                             "Sample operation '{}' must have either 'samples' or 'filePath'",
@@ -303,7 +308,9 @@ pub async fn op_playback_build_graph(
                         let samples = if let Some(ref samples) = input.samples {
                             samples.clone()
                         } else if let Some(ref file_path) = input.file_path {
-                            load_audio_samples(file_path, sample_rate, channels)?
+                            let path = std::path::PathBuf::from(file_path);
+                            let buffer = sample_cache.get_or_load(path, sample_rate, channels)?;
+                            buffer.data.clone()
                         } else {
                             return Err(format!(
                                 "Merge input {} in operation '{}' must have either 'samples' or 'filePath'",
@@ -1082,240 +1089,6 @@ fn stop_current_playback(state: &OpPlaybackState) {
         s.clear();
     }
     *sink = None;
-}
-
-fn load_audio_samples(
-    file_path: &str,
-    target_sample_rate: u32,
-    target_channels: u16,
-) -> Result<Vec<f32>, String> {
-    use std::fs::File;
-
-    // Use symphonia for audio decoding
-    use symphonia::core::audio::SampleBuffer;
-    use symphonia::core::codecs::DecoderOptions;
-    use symphonia::core::formats::FormatOptions;
-    use symphonia::core::io::MediaSourceStream;
-    use symphonia::core::meta::MetadataOptions;
-    use symphonia::core::probe::Hint;
-
-    let file = File::open(file_path).map_err(|e| format!("Failed to open file: {}", e))?;
-    let mss = MediaSourceStream::new(Box::new(file), Default::default());
-
-    let mut hint = Hint::new();
-    if let Some(ext) = std::path::Path::new(file_path).extension() {
-        hint.with_extension(&ext.to_string_lossy());
-    }
-
-    let probed = symphonia::default::get_probe()
-        .format(
-            &hint,
-            mss,
-            &FormatOptions::default(),
-            &MetadataOptions::default(),
-        )
-        .map_err(|e| format!("Failed to probe file: {}", e))?;
-
-    let mut format = probed.format;
-
-    let track = format
-        .tracks()
-        .iter()
-        .find(|t| t.codec_params.codec != symphonia::core::codecs::CODEC_TYPE_NULL)
-        .ok_or("No audio track found")?;
-
-    // Get source audio properties
-    let source_sample_rate = track
-        .codec_params
-        .sample_rate
-        .ok_or("No sample rate found in audio file")?;
-
-    let source_channels = track.codec_params.channels.map(|c| c.count()).unwrap_or(1) as u16;
-
-    let mut decoder = symphonia::default::get_codecs()
-        .make(&track.codec_params, &DecoderOptions::default())
-        .map_err(|e| format!("Failed to create decoder: {}", e))?;
-
-    let track_id = track.id;
-    let mut samples: Vec<f32> = Vec::new();
-
-    // Decode all samples first
-    loop {
-        let packet = match format.next_packet() {
-            Ok(packet) => packet,
-            Err(symphonia::core::errors::Error::IoError(ref e))
-                if e.kind() == std::io::ErrorKind::UnexpectedEof =>
-            {
-                break;
-            }
-            Err(e) => return Err(format!("Error reading packet: {}", e)),
-        };
-
-        if packet.track_id() != track_id {
-            continue;
-        }
-
-        let decoded = match decoder.decode(&packet) {
-            Ok(decoded) => decoded,
-            Err(e) => {
-                eprintln!("Error decoding packet: {}", e);
-                continue;
-            }
-        };
-
-        let mut sample_buf = SampleBuffer::<f32>::new(decoded.capacity() as u64, *decoded.spec());
-        sample_buf.copy_interleaved_ref(decoded);
-        samples.extend_from_slice(sample_buf.samples());
-    }
-
-    // Apply channel conversion if needed
-    let samples = if source_channels != target_channels {
-        eprintln!(
-            "Converting audio channels: {} -> {} channels for file: {}",
-            source_channels, target_channels, file_path
-        );
-        convert_channels(&samples, source_channels, target_channels)?
-    } else {
-        samples
-    };
-
-    // Apply sample rate conversion if needed
-    let samples = if source_sample_rate != target_sample_rate {
-        eprintln!(
-            "Resampling audio: {}Hz -> {}Hz for file: {}",
-            source_sample_rate, target_sample_rate, file_path
-        );
-        resample_audio(
-            &samples,
-            source_sample_rate,
-            target_sample_rate,
-            target_channels,
-        )?
-    } else {
-        samples
-    };
-
-    eprintln!(
-        "Loaded audio file: {} ({}Hz, {} channels, {} samples)",
-        file_path,
-        target_sample_rate,
-        target_channels,
-        samples.len()
-    );
-
-    Ok(samples)
-}
-
-/// Convert between different channel counts
-fn convert_channels(
-    samples: &[f32],
-    source_channels: u16,
-    target_channels: u16,
-) -> Result<Vec<f32>, String> {
-    if source_channels == target_channels {
-        return Ok(samples.to_vec());
-    }
-
-    let frames = samples.len() / source_channels as usize;
-    let mut output = Vec::with_capacity(frames * target_channels as usize);
-
-    for frame_idx in 0..frames {
-        let source_frame_start = frame_idx * source_channels as usize;
-
-        match (source_channels, target_channels) {
-            // Mono to Stereo: duplicate the mono channel
-            (1, 2) => {
-                let mono_sample = samples[source_frame_start];
-                output.push(mono_sample); // Left
-                output.push(mono_sample); // Right
-            }
-            // Stereo to Mono: average left and right channels
-            (2, 1) => {
-                let left = samples[source_frame_start];
-                let right = samples[source_frame_start + 1];
-                let mono = (left + right) * 0.5;
-                output.push(mono);
-            }
-            // Multi-channel to Stereo: downmix by averaging all channels
-            (src, 2) if src > 2 => {
-                let mut sum = 0.0;
-                for ch in 0..src {
-                    sum += samples[source_frame_start + ch as usize];
-                }
-                let avg = sum / src as f32;
-                output.push(avg); // Left
-                output.push(avg); // Right
-            }
-            // Multi-channel to Mono: downmix by averaging all channels
-            (src, 1) if src > 1 => {
-                let mut sum = 0.0;
-                for ch in 0..src {
-                    sum += samples[source_frame_start + ch as usize];
-                }
-                let avg = sum / src as f32;
-                output.push(avg);
-            }
-            // Unsupported conversion
-            (src, tgt) => {
-                return Err(format!(
-                    "Unsupported channel conversion: {} -> {} channels",
-                    src, tgt
-                ));
-            }
-        }
-    }
-
-    Ok(output)
-}
-
-/// Simple linear interpolation resampler
-fn resample_audio(
-    samples: &[f32],
-    source_rate: u32,
-    target_rate: u32,
-    channels: u16,
-) -> Result<Vec<f32>, String> {
-    if source_rate == target_rate {
-        return Ok(samples.to_vec());
-    }
-
-    let source_frames = samples.len() / channels as usize;
-    let ratio = target_rate as f64 / source_rate as f64;
-    let target_frames = (source_frames as f64 * ratio).ceil() as usize;
-
-    let mut output = Vec::with_capacity(target_frames * channels as usize);
-
-    for target_frame in 0..target_frames {
-        let source_pos = target_frame as f64 / ratio;
-        let source_frame = source_pos.floor() as usize;
-        let frac = source_pos - source_frame as f64;
-
-        for ch in 0..channels {
-            let ch_idx = ch as usize;
-
-            // Get current sample
-            let current_idx = source_frame * channels as usize + ch_idx;
-            let current_sample = if current_idx < samples.len() {
-                samples[current_idx]
-            } else {
-                0.0 // Pad with silence if beyond end
-            };
-
-            // Get next sample for interpolation
-            let next_idx = (source_frame + 1) * channels as usize + ch_idx;
-            let next_sample = if next_idx < samples.len() {
-                samples[next_idx]
-            } else {
-                current_sample // Use current if no next sample
-            };
-
-            // Linear interpolation
-            let interpolated = current_sample + (next_sample - current_sample) * frac as f32;
-            output.push(interpolated);
-        }
-    }
-
-    Ok(output)
 }
 
 fn count_graph_ops(request: &BuildGraphRequest) -> Result<usize, String> {
