+++ src-tauri/src/playback/mod.rs
@@ -1,7 +1,4 @@
-pub mod file_source;
+pub mod audio_source;
 pub mod op_playback;
-pub mod region_source;
-pub mod source;
-
+pub use audio_source::*;
 pub use op_playback::*;
-pub use source::*;
